requirements.txt
#streamlit==1.28.0
openai>=1.12.0
google-cloud-aiplatform>=1.43.0
google-auth==2.22.0
python-dotenv==1.0.0
SQLAlchemy==2.0.29
pdfminer.six==20221105
python-docx==0.8.11
pypdf==3.11.0
numpy==2.1.0
scikit-learn>=1.4.0
sendgrid==6.11.0
gunicorn==20.1.0
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.32.3
Faker>=19.0.0
# urllib3>=2.3.0

# MCP Server dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
